Updates references to classes moved during package rename.
