+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/event/queues/processors/ScanAtomProcessorTest.java
@@ -33,9 +33,9 @@ import org.eclipse.scanning.event.queues.QueueServicesHolder;
 import org.eclipse.scanning.event.queues.beans.ScanAtom;
 import org.eclipse.scanning.event.queues.processors.ScanAtomProcessor;
 import org.eclipse.scanning.points.serialization.PointsModelMarshaller;
-import org.eclipse.scanning.test.event.queues.beans.util.TestAtomMaker;
 import org.eclipse.scanning.test.event.queues.mocks.DummyQueueable;
 import org.eclipse.scanning.test.event.queues.mocks.MockPublisher;
+import org.eclipse.scanning.test.event.queues.util.TestAtomMaker;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
